setup redux and store; move user context into redux store
