+++ package.json
@@ -9,8 +9,11 @@
     "firebase": "^10.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.1.2",
     "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.1",
+    "redux": "^4.2.1",
+    "redux-logger": "^3.0.6",
     "styled-components": "^6.0.8",
     "web-vitals": "^2.1.4"
   },
