+++ src/App.js
@@ -1,15 +1,35 @@
+import { useEffect } from 'react';
 import { Routes, Route, Navigate } from 'react-router-dom';
-import { useContext } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 
 import Home from './routes/home/home.component';
 import Navigation from './routes/navigation/navigation.component';
-import { UserContext } from './contexts/user.context';
 import Authentication from './routes/authentication/authentication.component';
 import Shop from './routes/shop/shop.component';
 import Checkout from './routes/checkout/checkout.component';
 
+import { setCurrentUser } from './store/user/user.action';
+import { selectCurrentUser } from './store/user/user.selector';
+import {
+  onAuthStateChangedListener,
+  createUserDocumentFromAuth
+} from './utils/firebase/firebase.utils';
+
 const App = () => {
-  const { currentUser } = useContext(UserContext);
+  const currentUser = useSelector(selectCurrentUser);
+
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChangedListener((user) => {
+      if (user) {
+        createUserDocumentFromAuth(user);
+      }
+      dispatch(setCurrentUser(user));
+    });
+
+    return unsubscribe;
+  }, []);
 
   return (
     <Routes>
