+++ src/routes/navigation/navigation.component.jsx
@@ -1,17 +1,16 @@
 import { Fragment, useContext } from 'react';
 import { Outlet } from 'react-router-dom';
-
-import { ReactComponent as CrwnLogo } from '../../assets/crown.svg';
-
-import { UserContext } from '../../contexts/user.context';
-import { CartContext } from '../../contexts/cart.context';
-
-import { signOutUser } from '../../utils/firebase/firebase.utils';
+import { useSelector } from 'react-redux';
 
 import CartIcon from '../../components/cart-icon/cart-icon.component';
 import CartDropdown from '../../components/cart-dropdown/cart-dropdown.component';
 
-import './navigation.styles.jsx';
+import { CartContext } from '../../contexts/cart.context';
+import { selectCurrentUser } from '../../store/user/user.selector';
+
+import { ReactComponent as CrwnLogo } from '../../assets/crown.svg';
+import { signOutUser } from '../../utils/firebase/firebase.utils';
+
 import {
   LogoContainer,
   NavLink,
@@ -20,7 +19,8 @@ import {
 } from './navigation.styles.jsx';
 
 const Navigation = () => {
-  const { currentUser } = useContext(UserContext);
+  const currentUser = useSelector(selectCurrentUser);
+
   const { isCartOpen } = useContext(CartContext);
 
   return (
