+++ src/store/user/user.reducer.js
@@ -1,0 +1,19 @@
+import USER_ACTION_TYPES from './user.types';
+
+const INITIAL_STATE = {
+  currrentUser: null
+};
+
+export const userReducer = (state = INITIAL_STATE, action) => {
+  const { type, payload } = action;
+
+  switch (type) {
+    case USER_ACTION_TYPES.SET_CURRENT_USER:
+      return {
+        ...state,
+        currentUser: payload
+      };
+    default:
+      return state;
+  }
+};
