+++ src/store/user/user.types.js
@@ -1,0 +1,5 @@
+const USER_ACTION_TYPES = {
+  SET_CURRENT_USER: 'user/SET_CURRENT_USER'
+};
+
+export default USER_ACTION_TYPES;
